+++ internal/credentials/oauth2.go
@@ -143,16 +143,16 @@ func WithGrantType(grantType string) grantTypeOption {
 }
 
 // Resource
-type resourceOption string
+type resourceOption []string
 
 func (resource resourceOption) ApplyOauth2CredentialsOption(c *oauth2TokenExchange) error {
-	c.resource = string(resource)
+	c.resource = resource
 
 	return nil
 }
 
-func WithResource(resource string) resourceOption {
-	return resourceOption(resource)
+func WithResource(resource ...string) resourceOption {
+	return resource
 }
 
 // RequestedTokenType
@@ -309,7 +309,7 @@ type oauth2TokenExchange struct {
 	// urn:ietf:params:oauth:grant-type:token-exchange by default
 	grantType string
 
-	resource string
+	resource []string
 	audience []string
 	scope    []string
 
@@ -607,7 +607,7 @@ func (cfg *oauth2TokenSourceConfig) applyConfig(tokenSrcType int) (*tokenSourceO
 //nolint:tagliatelle
 type oauth2Config struct {
 	GrantType          string               `json:"grant-type"`
-	Resource           string               `json:"res"`
+	Resource           *stringOrArrayConfig `json:"res"`
 	Audience           *stringOrArrayConfig `json:"aud"`
 	Scope              *stringOrArrayConfig `json:"scope"`
 	RequestedTokenType string               `json:"requested-token-type"`
@@ -622,8 +622,8 @@ func (cfg *oauth2Config) applyConfig(opts *[]Oauth2TokenExchangeCredentialsOptio
 		*opts = append(*opts, WithGrantType(cfg.GrantType))
 	}
 
-	if cfg.Resource != "" {
-		*opts = append(*opts, WithResource(cfg.Resource))
+	if cfg.Resource != nil && len(cfg.Resource.Values) > 0 {
+		*opts = append(*opts, WithResource(cfg.Resource.Values...))
 	}
 
 	if cfg.Audience != nil && len(cfg.Audience.Values) > 0 {
@@ -723,8 +723,10 @@ func (provider *oauth2TokenExchange) addTokenSrc(params *url.Values, src TokenSo
 func (provider *oauth2TokenExchange) getRequestParams() (string, error) {
 	params := url.Values{}
 	params.Set("grant_type", provider.grantType)
-	if provider.resource != "" {
-		params.Set("resource", provider.resource)
+	for _, res := range provider.resource {
+		if res != "" {
+			params.Add("resource", res)
+		}
 	}
 	for _, aud := range provider.audience {
 		if aud != "" {
@@ -1054,7 +1056,7 @@ func (provider *oauth2TokenExchange) String() string {
 	defer buffer.Free()
 	fmt.Fprintf(
 		buffer,
-		"OAuth2TokenExchange{Endpoint:%q,GrantType:%s,Resource:%s,Audience:%v,Scope:%v,RequestedTokenType:%s",
+		"OAuth2TokenExchange{Endpoint:%q,GrantType:%s,Resource:%v,Audience:%v,Scope:%v,RequestedTokenType:%s",
 		provider.tokenEndpoint,
 		provider.grantType,
 		provider.resource,